requirements.txt
numpy==1.21.3
Pillow==9.0.0
